+++ azure-kusto-ingest/azure/kusto/ingest/__init__.py
@@ -1,4 +1,4 @@
-from .kusto_ingest_client import KustoIngestClient
+from ._ingest_client import KustoIngestClient
 from ._descriptors import BlobDescriptor, FileDescriptor
 from ._ingestion_properties import (
     DataFormat,
